refactor(scraper): struture of the python app
