+++ scraper/.docker/docker-entrypoint-initdb.d/create-appuser.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -e
 
 # a default non-root role
