--- scraper/isp-scrape.sh
@@ -1,4 +1,0 @@
-#!/bin/zsh
-
-pip3 install -r requirements.txt
-python3 ./DataCollector/speedtest_collector.py
